finnaly init 2.0

+++ app/routers/users.py
@@ -1,16 +1,19 @@
 import uuid
+from typing import List, Optional
 
 from fastapi import APIRouter, Depends, HTTPException
 from fastapi_cache import FastAPICache
 from fastapi_cache.decorator import cache
+from sqlalchemy import func
 from sqlalchemy.orm import Session
 from starlette import status
 
 from app import utils, models, oauth2
+from app.config import settings
 from app.database import get_db
 from app.oauth2 import is_current_user_admin, get_current_user
 from app.schemas import user_schemas
-from app.utils import validate
+from app.utils import validate, validate_list
 
 router = APIRouter(prefix='/users', tags=['Пользователи'])
 
@@ -22,6 +25,22 @@ async def get_user(current_user: models.User = Depends(get_current_user), db: Se
     return await validate(value=user, class_type=user_schemas.UserOut)
 
 
+@router.get('/', response_model=List[user_schemas.UserOut])
+@cache(namespace='users')
+async def get_users(current_user_admin=Depends(is_current_user_admin), db: Session = Depends(get_db)):
+    users = db.query(models.User).all()
+    return await validate_list(values=users, class_type=user_schemas.UserOut)
+
+
+@router.get('/search', response_model=List[user_schemas.UserAdminSearch])
+async def search_users(query: Optional[str], current_user_admin=Depends(is_current_user_admin),
+                       db: Session = Depends(get_db)):
+    users = db.query(models.User).filter(
+        func.lower(models.User.firstname + models.User.surname + models.User.email).contains(query.lower())).limit(
+        5).all()
+    return users
+
+
 @router.post('/', status_code=status.HTTP_201_CREATED)
 async def create_user(new_data: user_schemas.UserCreate,
                       db: Session = Depends(get_db)):
@@ -30,6 +49,8 @@ async def create_user(new_data: user_schemas.UserCreate,
     db.add(new_user)
     db.commit()
     db.refresh(new_user)
+
+    await FastAPICache.clear(namespace='users')
 
     access_token = await oauth2.create_access_token(data={"user_id": str(new_user.id), "email": new_user.email})
     return {"access_token": access_token, "token_type": "bearer"}
@@ -43,7 +64,9 @@ async def update_user(id: uuid.UUID, updated_data: user_schemas.UserUpdate,
 
     if user is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Пользователь с ID {id} не найден.")
-    if updated_data['password']:
+    if user.email == settings.owner_email and current_user_admin.email != settings.owner_email:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Че брат, думал все так просто?")
+    if updated_data.password is not None:
         updated_data.password = utils.hash(updated_data.password)
 
     update_data = updated_data.dict(exclude_unset=True)
@@ -55,6 +78,7 @@ async def update_user(id: uuid.UUID, updated_data: user_schemas.UserUpdate,
         db.refresh(user)
 
         await FastAPICache.clear(namespace='users/me')
+        await FastAPICache.clear(namespace='users')
 
     return user
 
@@ -65,6 +89,8 @@ async def delete_user(id: uuid.UUID, current_user_admin=Depends(is_current_user_
     user = db.query(models.User).filter(models.User.id == id).first()
     if user is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Пользователь с ID {id} не найден.")
+    if user.email == settings.owner_email and current_user_admin.email != settings.owner_email:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Че брат, думал все так просто?")
     db.delete(user)
     db.commit()
